Finalised loading from a file andthe mechanism to update the map
Game logic now handles the input of save file there is a small bug in pass_object_values function on the pass of ids from the two buffer arrays to the game struct arrays. Also several new functions added one important one **update_objects()** every time this one is called checks if chests and monsters are opened or dead correspondingly, when a monster dies or a chest gets openned please set the value of their variables to TRUE. Several task functions added used to calculate variables etc most of the changes are in gamelogic. If you choose to debug to solve the bug referenced above try with gdb. I'll try to solve this later today.To connect from a saved game use the following credentials whoami:root otherwise since you'll get connected to a new game since theresn't any save game for that account yet. Need a way to show that either a new game or a saved game is loading