Fixed a major bug, implemented save game
DISCLAIMER: There is one bug left to be taken down, that's the number of chests on load. I also implemented the save game and added it to the menu. If you want to test other functionalities such as battle system please comment them out as I said there's still a bug that must be taken down since the load-save functionality is fully ready to deploy.